Implement field of vision
This commit add field of vision, implemented using raycasting with Bresenham's
line algorithm (not efficient but intuitive).

Some adjustments to existing code have been done along the way. The most
significant change might be, that the position of a tile (Tile:get[XY]) is now
guaranteed to be an integer.